Implement different stack size
Added a parameter howMany to the GenerateSpawnOrder() function to specify the stack size for each fruit.
Modified the GenerateSpawnOrders() function to handle different stack sizes and bonus fruit spawning .
If the randomly selected fruit is a bonus fruit, only one fruit is added to the spawn order stack.
For other fruits, the specified howMany value is used to determine the stack size.
Updated the GenerateSpawnOrder() function calls in the GenerateSpawnOrders() function to pass the stack size parameter.
This change allows for the spawning of bonus fruits and different stack sizes, enhancing the gameplay dynamics.